shell: Superuser init for remote hosts
When logging into a remote host with SSH, we now control whether a
superuser bridge is started up based on what has been done in the last
session, just like for the local login.
